requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client for API calls
httpx==0.25.2
requests==2.31.0

# Vector Database Plugins
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone==5.0.0

# LLM API Plugins
openai==1.3.0
anthropic==0.7.0

# Document Processing Plugins
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2

# Embedding Models
sentence-transformers==2.2.2

# Database Connectors
# sqlite3 is built into Python
psycopg2-binary==2.9.9  # PostgreSQL
pymysql==1.1.0  # MySQL

# Utility Libraries
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1